fix(2660): address Copilot review — livelock + pinned-anchor bypass + cleanups
Closes the 9 review comments on PR #2660 in one pass.

Critical (real bugs):
- consensus/network.rs: Bootstrapping→Idle gate evaluated only on the
  current_bft_mode != last_bft_mode edge, so a node entering BFT mode
  before any peer height was observed (last_height_seen == 0) stayed in
  Bootstrapping forever. Moved the gate out of the transition block — it
  now runs every tick whenever (BFT mode active AND state ==
  Bootstrapping), so peer heights arriving later still unblock us.
- network/mod.rs `state` field: `identity_not_registered` arm was
  unreachable because the earlier `chain_height == 0 → "connecting"`
  branch returned first. Reordered so the more-specific
  identity-not-registered arm takes precedence.
- web4/trust.rs: bootstrap-mode SPKI rotation branch fired on any
  existing anchor — including `TrustPolicy::Pinned`. Guard with
  `!matches!(policy, Pinned)` so explicit operator pins still reject on
  mismatch, matching the comment that said only TOFU anchors rotate.

CI guard:
- check-data-paths.sh regex caught only the raw literal and
  `PathBuf::from`. Extended to `(PathBuf|Path)::(from|from_str|new)`
  with optional `std::path::` qualifier so `Path::new("./data/...")` and
  fully-qualified forms also trip the guard.

Test flakiness:
- quic_mesh.rs test ctor and lib-network test_utils used
  `temp_dir.join("zhtp-test-tls-{process::id()}")`. Cargo runs tests in
  parallel within one process so concurrent calls shared TLS files.
  Added per-call counter + nanosecond suffix.

Cleanups:
- token/mod.rs: dropped two redundant `as u128` casts on
  `get_token_balance(...).unwrap_or(0)` — already u128.
- zhtp/src/lib.rs: added `debug_assert!(subpath.is_relative())` to
  `node_data_path` so an accidental absolute subpath (which `join`
  silently routes outside the data dir) panics in debug builds.

+++ zhtp/src/api/handlers/token/mod.rs
@@ -558,9 +558,10 @@ impl TokenHandler {
             if let Some(store) = blockchain.get_store() {
                 let storage_token_id = lib_blockchain::storage::TokenId(token_id_array);
                 let addr = lib_blockchain::storage::Address::new(target_key_id);
+                // get_token_balance returns Amount (u128) — no cast needed (CR #2660).
                 store
                     .get_token_balance(&storage_token_id, &addr)
-                    .unwrap_or(0) as u128
+                    .unwrap_or(0)
             } else {
                 token
                     .find_balance_by_key_id(&target_key_id)
@@ -741,9 +742,10 @@ impl TokenHandler {
                 if let Some(store) = blockchain.get_store() {
                     let storage_token_id = lib_blockchain::storage::TokenId(*token_id);
                     let addr = lib_blockchain::storage::Address::new(target_key_id);
+                    // get_token_balance returns Amount (u128) — no cast needed (CR #2660).
                     store
                         .get_token_balance(&storage_token_id, &addr)
-                        .unwrap_or(0) as u128
+                        .unwrap_or(0)
                 } else {
                     token
                         .find_balance_by_key_id(&target_key_id)
